Introduce prototype for upload API, based on S3
This API can be used to upload objects.
Rather than inventing our own API for this, let's use a subset of
the true S3 API. This allows usage of existing clients (such as
boto) to upload content through exodus-gw. The examples/s3-upload
script can be used to demonstrate this.

This implementation is prototype-level and comes with a TODO list.
Some major missing pieces are:

- performance testing to verify that this doesn't perform notably
  worse than using S3 directly
- a proper way to handle configuration (it currently allows the boto
  library to use whatever config is discovered by default, meaning
  only a single environment is supported)

This commit also experiments a bit more on how to document these
APIs.  There is now a separate API user guide and reference, as it's
not clear if there's a decent way of including general usage info
into the redoc-hosted reference.