deploy: fix nested actor invocation & use real broker in tests
The usage of send_with_options was not correct as the kwargs for the
actor need to be put into a 'kwargs' structure. Also, task.id is a UUID
and it's not possible to include that in the message currently as it
won't serialize to JSON.

Fix those issues and change the relevant test to use a real broker
rather than a mock. Using a real broker makes the test more accurate,
and at least allows catching the second problem above (though not the
first unfortunately).