docs: allow for customization of the application's main doc string
The built-in docs are worded in a generic way; e.g. since authentication
is beyond the scope of exodus-gw itself, the doc string under that
section vaguely suggests to consult "internal documentation".

Internal users need a way to actually find the docs relevant to them, so
let's allow these sections to be customized. Let the doc strings use env
var substitution. This means we can set an environment variable to link
to proper internal docs about authentication, or to explain that the
specific deployed environments are "pre" and "live", etc.